Provide access to rate_reset, a response header
Foursquare API returns a 403 error, and the response object returned by their API will be empty, if user is over their quota limits. They will also include a X-RateLimit-Reset header in the response, which is a timestamp that corresponds to when the rate limits will reset. This info can be found here: https://developer.foursquare.com/docs/places-api/rate-limits/. The proposed changes are a start, there may be additional exception handling that needs to be done to fully implement this change. However, I think the author of this library may know better where changes need to be made. However, for this library to be useful for users with quotas, I think this change is necessary. Feel free to add to, make comments, and build off of this proposal.